Kill preexisting when ibazel run detects changes
It feels silly that I didn't catch this sooner cause it is obvious in
hindsight. When you run `bazel run` on a task it blocks usage of Bazel
until the task finishes. Normally this isn't a problem, but the most
useful use of `bazel run` is when working with servers that are long
lived and need to be restarted. Since Bazel was locked by the naive
`bazel run` call, ibazel was never able to get the lock and actually run
the new code thus completely defeting the purpose of ibazel.

Now, we run `bazel build` and then start the subprocess, tracking the
PGID, and send a kill to the PGID when a change has been detected and a
rebuild is going to occur. This should make `ibazel run` useful in the
run mode.