Notify of changes command: Correctly identify when a build completes
(fixes #56) Prior to this change everytime ibazel rebuilt a bazel command it
would think that the build failed even if it succeeded.
The issue was the the underlying command object was being
called with both Run()# On branch master and subsequently Wait().
Becuase the command had already been run,
the invocation of Wait() always produced an error.