Warn when using globally installed npm package
Followup to #74
When we roll out the feature of searching for a locally installed ibazel, we need to communicate it to users,
or they are not likely to discover it.
Especially while ibazel is under active development, development workflows like browser livereload are
sensitive to the ibazel version so I want to be sure users are not likely to trip on versioning.