Move e2e output assertions into helper
This allows much longer timeouts to be used while still achieving faster
runtimes. The way I do this is by moving the assertion from the test
intointo the e2e helper. This way the call to get the output knows what
it is waiting for. Since it knows what it is waiting for, if it gets it
before time is up then it's free to continue onto the next step. If it
doesn't get it then it will wait until the timeout arrives.