Add test reproducing vim backupcopy=no and kqueue failure (#139)
* Add test reproducing vim backupcopy=no and kqueue failure

This test reproduces the write behavior of vim with `backupcopy` set to
both `yes` and `no`. `yes` is the default on Linux, and #10 fixed this
case. Unforunately, on macOS, `yes` is not the default, and using `no`
will cause ibazel to stop watching the file after a single file save
when a kqueue-based file watcher is used.

* Readd files when they are moved to overcome limitations in kqueue.