move java sources under brs/.
at head, bazel running a serve() target (e.g. //examples:simple)
explodes when the executable brs is not found. this appears to be a bug
with targets at the root level. instead of investigating this, just move
the java files to a subpackage (which i kind of wanted to do anyway).