Enhance the live_reload test
Move the test into golang as the test language. This has the benefit of
making it a lot easier to test the signaling that is also functioning in
a test.

Since it's in golang, why not include an example client that is compiled
inside the test and can be referred to in documentation.