Add support for command execution after successful execution
The "-run_command_after_success" flag is nearly identical to
the "-run_command_after" flag, but instead of operating
unconditionally, it only executes if the previous command
was successful.

This can provide value for exeucting a command which operates
on the output of a bazel build, and which should only
execute if there is new output to act upon.