Add flag to force-enable JS serving
For unclear reasons, `ibazel` doesn't detect our use of the
`ibazel_live_reload` tag, and therefore never starts the live
reload server. Fixes and closes bazelbuild/bazel-watcher#358.

When our server loads up and runs, it never sees the environment
variable. If we try to load the `lrserver`'s `livereload.js` at
the default port, it fails -- because the server never starts
listening in the first place.

This adds a simple flag to `ibazel`: `force_live_reload`. When this
flag is passed, `ibazel` starts up the `lrserver` in `Initialize`
instead of during target scanning.