Update directory structure
Since 2017 I've learned a lot more about how to structure go projects.
This is a big refactor to move the things I don't intend to be imported
by anyone else into an `internal` directory which will prevent external
parties from importing from it. This gives freedom to manipulate and
make backward-incompatible API changes inside the `internal` directory
going forward. Sorry if this breaks you, but it was never the intent of
the project to create an API that provides programmatic control of
`bazel`'s CLI. This is a breaking change but not one that I'm going to
be rolling back because I don't have bandwidth to support that usecase.
If you're using it as such I would recommend going back to the parent
commit of this one, copying the directory you need into your project and
working with it from there.