notify: `Wait` on subprocess so that `ProcessState` is populated
In the Notify code path bazel-watcher uses `os.ProcessState` to determine if a process
is still running. However, it never calls `Wait` or `Run`, like
it does in the normal codepath. This patch adds an async goroutine that `Wait`s for the
subprocess to exisit so that `ProcessState` is populated correctly.

Without this change subprocesses that exit result in bazel-watcher never detecting that
the process is dead and trying to write to it continuously, which causes broken pipe
errors in macOS/Linux.